requirements.txt
Package                Version
---------------------- ---------
certifi                2022.12.7
charset-normalizer     3.0.1
click                  8.1.3
colorama               0.4.6
contourpy              1.0.7
cycler                 0.11.0
Flask                  2.2.2
fonttools              4.38.0
idna                   3.4
itsdangerous           2.1.2
Jinja2                 3.1.2
kiwisolver             1.4.4
MarkupSafe             2.1.2
matplotlib             3.6.3
mysql-connector        2.2.9
mysql-connector-python 8.0.32
numpy                  1.24.1
packaging              23.0
pandas                 1.5.3
Pillow                 9.4.0
pip                    22.2.1
protobuf               3.20.3
pyparsing              3.0.9
python-dateutil        2.8.2
pytz                   2022.7.1
requests               2.28.2
setuptools             63.2.0
six                    1.16.0
urllib3                1.26.14
Werkzeug               2.2.2
